refactor: rename libs
License: MIT
